For recorders that are common across preinstall and postinstall (most of them) - share them properly
