+++ src/recorders/mod.rs
@@ -1,2 +1,3 @@
+pub mod common;
 pub mod postinstall;
 pub mod preinstall;
